requirements.txt
flask>=2.3.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
Werkzeug>=2.3.0
